Changed unit value from conversion rate to value in base units.
The unit values returned from the unit repository used to represent the
conversion rate compared to the base unit. Now the unit value is the
number of base units represented by the unit. For example; if the base
unit is 1 meter, then the unit value of millimeters is 0,001, because 1
millimeter is 0,001 meter.